Foncrion servos en mode non bloquant, refaire les courbes d'ir

+++ POC/MPLAB/marvin_v02.X/Servo.h
@@ -11,8 +11,8 @@
 #include "pwm.h"
 
 void    marvin_attach_servo(m_servo *servo, u32 *pin ,u32 *ocrs, u16 min, u16 max, u8 oc_timer, u16 periode);
-void    marvin_move_servo(m_servo *servo, u8 angle);
-void    marvin_move_servo_speed(m_servo *servo, u8 angle, u16 deg_per_periode, u16 periode_msec);
+void    marvin_move_servo(m_servo *servo, s16 angle);
+void    marvin_move_servo_speed(m_servo *servo, s16 angle, u16 deg_per_periode, u16 periode_msec);
 
 #endif	/* SERVO_H */
